client: Don't add dependencies we've already seen
This fixes an issue where infinite recursion could happen when
a module transitively dependended on itself. The public_deps
dict is apparently supposed to avoid that, but it seems like it
only works for modules that have an existing conan_ref, which might
not be true of a module at the root of the tree.

This adds a new set argument to _load_deps which ensures that
dependencies with the same name are not loaded again.