Add include paths to CFLAGS and CPPFLAGS.
Some configure scripts doesn't seem to honor C_INLUDE_PATH and
CPP_INCLUDE_PATH but instead expect -I <path> to be set in CFLAGS
and CPPFLAGS. I ran into this in the Python configure script

Adding them here should not have any negative effects except listing
some of the include paths twice if the configure script honors
both environments.

An alternative approach would be to allow a option to be set to
tell it where to add the include paths, but it might be to complicated.