Fix tests on recent versions of OSX and Xcode
The settings in the update_settings_yml_test failed because
my compiler was a higher version number than the ones listed
in there. I have updated the settings.yml in the source file
but we might want to make sure that these settings are shared
between the test and what we ship in the acutal application
to avoid these problems in the future.