Quote string(REPLACE) cmake command fourth argument
This commit fixes potential errors when CONAN_VERSION variable
in conan_split_version() function is not set so the command
is interpreted as with 3 arguments only. This could also help if
the variable has spaces so the command is interpreted with 5 arguments.
As a rule, always quote cmake command arguments