Move graphing logic into `conan info`
Prior to this commit, the graphing logic was contained within its own
`conan graph` command, but this resulted in a lot of overlapping code
with the `conan info` command. This commit changes the logic to instead
be inside of a new argument to `conan info` called with `--graph` or
`-g`.