Add include directories for the Resource compiler
visual_studio generator: Add include directories for the resource compiler, so that it can find and use *.h files that define resource symbols and IDs. This will allow dependencies to use resources from the projects they depend on. (use the same directories as normal header includes)

This change should not break anyone because they will only need it if they are including header files, and if they are doing that, the header files should have custom names anyway.

Related to conan-io #1421.