Update files.py (#1793)
A few of us felt this error was a bit misleading with the capital L, knowing that linux is case sensitive, some wondered if perhaps this meant it was looking for a `Lib` folder with a capital `L` .  Seeing that is can be passed in as a parameter, it doesn't feel like a good improvement to change the error to have a lower case `l` which would still be confusing if you passed in a non-default folder name.  So, this seemed more robust, although it might make the message too long.  Just a simple suggestion, feel free to reject.