installer: add support to user-provided copytree function
On Windows, shutil.copytree performs rather poorly. There are faster
options for copying large amount of files that will use the hardware
more effectively, such as robocopy and teracopy. Rather than supporting
an individual tool, I think it's best to allow users to provide a
shutil.copytree-like function as a ConanFile member which will be
picked up and used automatically.