Add support for checking GPG signature of a file
Add a function in client tools which uses pygnupg library to check
GPG signatures of files. Useful for many open source projects which
distribute signed release tarballs.

A monkey-patch to pygnupg is required for the case of binary files in
python3. This has been submitted upstream, but since it is a very simple
change, the patch is dynamically applied in this file as well.

Signed-off-by: Dan Weatherill <plasteredparrot@gmail.com>