Workspace: Use CMAKE_BINARY_DIR to refer to installation folder
This fixes up a problem when using 'conan install' to create multiple build
directories, for instance, to create both a cmake-build-debug and a
cmake-build-release directory for use with CLion.

The CMakeLists.txt is repeatedly generated in that case, and the binary_dir of
each subdirectory points to the last build directory created. It is amazing what
number of strange CMake errors arise when the directories are intertwined.

Instead, use CMAKE_BINARY_DIR, which points to the build directory in use.