Fix cmake_paths transitive dependencies #4467
This commit fixes an issue when using the cmake_paths generator with
packages that itself have one or more transitive package dependencies.
Currently only the CMake variables CMAKE_MODULE_PATH and
CMAKE_PREFIX_PATH get populated in the conan_paths.cmake toolchain file.

However, this information is not sufficient for packages that have been
generated using the cmake generator as they may use Conan specific
variables in the form of CONAN_*_ROOT that are populated with the paths
pointing to its own, transitive dependencies.

Since this information is already stored in |DepsCppCmake| when creating
the conan_paths.cmake toolchain file, this commit just populates the custom
Conan variables in this file as well.