Update conan_requester.py
Hi, I have a problem where proxy settings are _not_ being taken from the environment (so I have to manually specify them). In searching for the solution to that problem I noticed that this code might not work as intended. When the popped variable is True the right part of the "or" will not be executed ... and thus not all proxy settings would be cleared (probably only the first encountered).

>>> False or print("hi")
hi
>>> True or print("hi")
True  <=== the right part of the or did not execute.

Do I understand it right? Thanks.