Profile include order (#6398)
* Fix profile include order variable overwriting

If you had a setup like this:

myinclude
---------
MYVAR=foo
---------

myprofile
---------
include(myinclude)
MYVAR=bar
---------

The end result would be MYVAR=foo which is *very* confusing
behavior. I think most users would expect MYVAR to be bar.

This was caused by inherited_vars being a reference and not
a copy. Explicity copying the profile_parser.vars fixes the
problem which leads me to think this is the original intent
of this function since it matches what my expection would be
as well.

* fixing evaluation too

Co-authored-by: James <james@conan.io>