Don't suppress all 'os.makedirs' exceptions.
os.makedirs can raise a number of exceptions, one of which (already
exists) could safely be ignored while we were silencing all possible
exceptions.

As an example of an issue that this could cause, if the ".conan"
directory was located in a directory the user doesn't have write
permission to, conans.util.files.save would raise a FileNotFoundError,
which is misleading.

Fixes #7281.