Pass dependency version to CMake layer in conanbuildinfo.cmake
This is useful to run configure_file() on a conanfile.txt.in (as part of
an example) to automatically have the dependency versions updated.
Other usecases might include version dependent switches in the CMake
layer.

Since the information is available in cpp_info.version anyways, it is
straightforward to pass it on the CMake.