Could not apply chmod on Windows
There is a note on chmod docs:
https://docs.python.org/2.7/library/os.html#os.chmod

"Although Windows supports chmod(), you can only set the file’s
read-only flag with it (via the stat.S_IWRITE and stat.S_IREAD constants
or a corresponding integer value). All other bits are ignored."

However, it works for files, not for folder.

Signed-off-by: Uilian Ries <uilianries@gmail.com>