fix env_diff() on linux
Previously the 'export' command was used to print environment variables
in the env_diff() function. This output was then parsed, but this
parsing code did not work for the output of the 'export' command.
'export' adds a prefix to each line, which was not accounted for.
By using the 'printenv' command on Linux instead no modification to the
parsing code is necessary and can be kept the same for Linux and Windows.