Recursively search 'requires' fields in lockfiles case of overridden dependencies
If a ConanFile defines a requirement with the property 'override=True',
GraphLock::lock_node() now searches the whole dependency graph instead
of just the direct dependencies of corresponding node.  This is
necessary because requirements of a ConanFile that were defined with the
property 'override=True' are not included in the 'requires' field of the
node that corresponds to that recipe.