Expand user-agent string to include OS info. (#8947)
It's useful to have the operating system information in the user-agent of the client to provide better analytics. Such analytics can help in determining if some particular set of clients are having problems. This adds the operating system, version, and machine type. In a format similar to what regular browsers provide.