Follow SONAME to generate BazelDeps correctly
If A and B are both shared libraries, and A links to B, the linker assigns the
NEEDED tag of A according to the SONAME tag of B. The file name of B was
chosen if and only if the SONAME tag of B does not exist.

Because of this rule, the BazelDeps should generate cc_import directive of
shared libraries from their SONAME tags instead of thier file names. Without
that, we meet library loading errors in runtime because the Bazel don't know
how to put the file (perhaps a symlink) named by its SONAME tag into the cache
directory.