xcodedeps: Fix dependency resolution if transitive dependencies have components
When a dependency of a dependency would have components, only the default
name "conan_dep_dep.xconfig` would be included. However, this file was never
generated, as they are in the form `conan_dep_component.xconfig`.

Fix this by adding all components instead.

Also allow the components to have dashes and use _format_name() to
avoid issues when parsing the xcconfig file.