Respect `force_host_context` when resolving dependencies
When a depedency is added with `test_requires` within `build_requirements`
we set `force_host_context` on the requirement.

However, when we build the graph and resolve the recursive dependencies,
we fail to respect that and set the profile/context to the target node
context.

If the original target node has a `build` context, we start to resolve
the nested node as `build` rather than as `host` context.

This patch adds a new branch to the context decision to respect the
`force_host_context` on the requirement and set the context appropriately.