test: Add xfail on the failing test
The test is only failing in the case of the default branch on the system
running the test is other than "master". To simulate this, you must
configure your global git to have, say, "main" as the default branch:

    git config --global init.defaultBranch main

After this, you can simply run this test case:

    python -m tox -- python -m tox -- conans/test/unittests/tools/scm/test_git_get_commit.py

The xfail will make it an expected failure.