fix(test): Create "master" branch after repo init
When initialising a new repository, Git will consider the
init.defaultBranch configuration setting that can be something other
than "master".

The git init command however accepts an option called "--init-branch"
which we can use to set the default branch explicitly. That solution,
however, can only work with newer Git versions.

The more versatile solution is to create the master branch if it does
not exist.

This way, whatever is the setting on the developer machine, the test
will pass.

Another solution could be to get the value of the configuration like so:

    git config --get init.defaultBranch

And use the output of the command above to checkout the default branch.
However, it would take a more code to pull this one out so I've decided
that the explicit setting of the default branch in the init call is the
most economical solution.