refactor: Replace deprecated imp.load_source with importlib
In Python 3.12 the imp module will be officially removed as it was
deprecated since 3.3. The test run already exposes a warning about this
and this can break as soon as we switch to Python 3.12 which is already
in alpha phase.

The code is more complex because the imp.load_source function is now
decomposed in a more flexible, powerful import customisation system but
that comes with a more verbose usage too.

The key here is to set up a loader that brings in the arbitrary file as
a python module code and then we need to create the module from the spec
of the loaded code. After that's done, we also have to execute the
module in order to have all the contents of the module available.