Fix test `git_get_commit` (#13105)
* test: Add xfail on the failing test

The test is only failing in the case of the default branch on the system
running the test is other than "master". To simulate this, you must
configure your global git to have, say, "main" as the default branch:

    git config --global init.defaultBranch main

After this, you can simply run this test case:

    python -m tox -- python -m tox -- conans/test/unittests/tools/scm/test_git_get_commit.py

The xfail will make it an expected failure.

* fix(test): Create "master" branch after repo init

When initialising a new repository, Git will consider the
init.defaultBranch configuration setting that can be something other
than "master".

The git init command however accepts an option called "--init-branch"
which we can use to set the default branch explicitly. That solution,
however, can only work with newer Git versions.

The more versatile solution is to create the master branch if it does
not exist.

This way, whatever is the setting on the developer machine, the test
will pass.

Another solution could be to get the value of the configuration like so:

    git config --get init.defaultBranch

And use the output of the command above to checkout the default branch.
However, it would take a more code to pull this one out so I've decided
that the explicit setting of the default branch in the init call is the
most economical solution.