Fixed parsing the output of vswhere
The output of vswhere normally uses "Ansi" encoding.
It was though decoded via "Utf-8", which raised
an exception in certain cases (i.e. letters like 'ö').

This commit forces the output to be utf-8, by providing the "utf8"
option to vswhere.

Thus, some visual studio installation could not be found.

Fixes #13281