Allow settings_target across compatibility method
In the info.py module, following the guidance provided in the comments
for the settings_target property, this commit focuses on refining the
handling of settings_target within various package methods.

When populating the settings_target property via the package_id()
method, the adjustment guarantees that it remains accessible to both the
info and original_info methods. By extending its accessibility to these
methods, it allows for seamless utilization within the compatibility()
method as well.

Signed-off-by: Adnan Ali <m_adnanali_1@hotmail.com>