bug fix: migrations do not respect core.cache:storage_path
migrations do not respect core.cache:storage_path.  If the user
has this option set, and tries to migrate from conan 2.0.13 to
2.0.14, the migration will fail.

This fix still assumes a 1-1 relationship between the $CONAN_HOME and
the local cache, but the forward and backward migrations are updated
to find the local cache in the correct way.