unset env ORIGINAL_PATH, so Conan's PATH will be used in msys
msys's /etc/profile checks and uses ORIGINAL_PATH as its PATH, instead
of the PATH in the environment.
We want to avoid this, and use the PATH that conan has prepared.

The design is so any msys shell started within an existing msys shell
will start with the original environment the first msys started with.

This applies when the user runs conan from within a msys terminal,
and builds a package that uses an embedded msys environment
(eg AutoTools-based packages).