Add properties to MakeDeps generator (#16613)
* Add properties to MakeDeps generator

If the receipt defines custom properties in the package_info() method
then these properties should appear as make variables, too. The pattern
of the make variable is

CONAN_PROPERTY_{dependency_name}_{property_name}

* Add property feature for components

Furthermore handle case when _properties is None.

* Extracting property makefication

* Add some tests

* Review finding

- _makefy_properties() always returns a dict
- Simplification of define_multiple_variable_value jinja2 macro

* Remove unnecessary extra space

Co-authored-by: James <memsharded@gmail.com>

* Skip properties with newlines

It would break the makefile if we would create property values with
newlines. Therefore we output a warning message and skip such
properties.

* Display scope of warning

---------

Co-authored-by: James <memsharded@gmail.com>