ISSUE-16731: "msvc" compatibility with "Visual Studio" is broken (#16732)
In case cppstd is set in the Conan profile for "Visual Studio", it is badly
migrated to "msvc". This change solves: https://github.com/conan-io/conan/issues/16731

The change also takes care of to handle the case when the 'default' cppstd is
set, in this case, it will be erased from the compatible package as well.