Report remote while requesting login
In setups with multiple remotes configured the prompt to enter a
password might appear multiple times. After the first input users could
deduce that the given password is wrong and in worst case start trying
other ones, even if it was correct. Naming the remote name in the prompt
as well will avoid such kind of confusion.