Fix Apt dpkg query arch when host_package=False
Apt.check(<pkg>, host_package=False) added the host arch,
instead of the build arch to the dpkg query, if used in
a recipe during cross-building conan install invocations.
This fix ensures that the `settings_build.arch` is used
to make the query in this context.

Issue: https://github.com/conan-io/conan/issues/19073